LayoutTests/http/tests/security/local-JavaScript-from-remote-expected.txt

CONSOLE MESSAGE: line 0: Not allowed to load local resource: file:///tmp/LayoutTests/http/tests/security/resources/localScript.js
This test is to see if a remote file can run a local script.

Test Passed.
